Fix credibility tests to work with Py3
To redirect stdout, different imports were needed for Py2 and Py3;
instead of execfile we work with mock argparse object.
Changed deprecated AssertEquals to AssertEqual in tests.
Closed files in tests because Py3 was producing a ResourceWarning.
Changed hyphen_if_empty to get_adapted_header_line for the sake of
readability.
Fixed a typo in analysis all_commands.