Add "within-phase" rules
Added rules that are specific within a single phase (sth has to be
unique within a phase etc.), moved checking for empty phases here to
reduce number of traversals over the parsed_input_list.
Fixed checking for journal end in single rules (was testing line instead
of command).
Solved pretty printing of output using an iterator to have a newline
after unknown flags and to print that there were no errors if none of
the present ones are enabled.
Fixed a typo.